fix: address multiple code quality and performance issues
This commit addresses several issues found during code audit:

**Security Fixes:**
- Add JSON validation type guards to LayoutContext and BookmarksPanel
  - Prevents prototype pollution via localStorage
  - Validates data structure and types before use
  - Uses fail-closed approach for unexpected properties

**Performance Optimizations:**
- Cache getBoundingClientRect calls in Timeline component
- Merge frame iteration loops in StreamDataContext (O(3n) → O(n))
- Fix unnecessary clones in index_session.rs using move semantics
- Debounce localStorage writes (BookmarksPanel, TooltipManager)
- Debounce thumbnail loading to reduce backend calls

**Code Quality:**
- Replace 65+ .lock().unwrap() calls with proper error handling macro
- Add bounds checking for frame indices in analysis commands
- Add symlink validation to path traversal checks
- Fix forEach early exit in keyboardShortcuts
- Replace magic numbers with constants (tooltip delays, z-index, offsets)
- Fix integer overflow in file size calculations

**New Helper Methods:**
- Add FrameType helper methods (is_key, is_inter, is_intra, etc.)
- Add short_name, from_str, description, and all() methods to FrameType

Files modified:
- crates/bitvue-core/src/types.rs: Add FrameType helper methods
- src-tauri/src/commands/file.rs: Fix integer overflow, add symlink validation
- src-tauri/src/services/decode_service.rs: Replace unwrap() with lock_mutex! macro
- src/components/BookmarksPanel.tsx: Add debounced localStorage saves
- src/components/Timeline.tsx: Cache getBoundingClientRect results
- src/constants/ui.ts: Add TOOLTIP_AUTO_HIDE_DELAY, Z_INDEX, DIMENSIONS constants
- src/contexts/StreamDataContext.tsx: Merge frame iteration loops
- src/contexts/ThumbnailContext.tsx: Add debounced thumbnail loading
- src/utils/interactiveTooltips/TooltipManager.ts: Use constants, add debouncing

+++ src-tauri/src/services/decode_service.rs
@@ -84,7 +84,8 @@ impl DecodeService {
             yuv_frames_cache: Mutex::new(HashMap::new()),
             rgb_lru_order: Mutex::new(VecDeque::new()),
             yuv_lru_order: Mutex::new(VecDeque::new()),
-            max_cache_bytes: 512 * 1024 * 1024, // 512MB default cache limit
+            // SAFETY: Use usize cast to prevent integer overflow (512 * 1024 * 1024 = 512MB)
+            max_cache_bytes: (512 * 1024 * 1024) as usize, // 512MB default cache limit
             current_cache_bytes: Mutex::new(0),
             mp4_samples_cache: Mutex::new(None),
             mkv_samples_cache: Mutex::new(None),
